+++ docs/changelog.md
@@ -11,7 +11,7 @@ Click on the version in the following list to see all changes since the previous
 
 ### [Unreleased]
 
-* ...
+* docs: change order of highlights in readme
 
 ### [3.12.0] - 2019-04-05