+++ docs/changelog.md
@@ -14,6 +14,7 @@ Click on the version in the following list to see all changes since the previous
 * docs: change order of highlights in readme
 * feat: migrate old votes with question mark to new string format
 * test: improve test coverage and suppress unused warnings
+* test: include performance test suite for Jira Data Center
 
 ### [3.12.0] - 2019-04-05