+++ docs/changelog.md
@@ -18,6 +18,7 @@ Click on the version in the following list to see all changes since the previous
 * test: add markers to session list and detail view to identify pages
 * feat: copy session link with a button on the share session screen (#62)
 * feat: document Jira Data Center compatibility in plugin descriptor
+* style: format css code consistently
 
 ### [3.12.0] - 2019-04-05