+++ docs/changelog.md
@@ -19,6 +19,7 @@ Click on the version in the following list to see all changes since the previous
 * feat: copy session link with a button on the share session screen (#62)
 * feat: document Jira Data Center compatibility in plugin descriptor
 * style: format css code consistently
+* feat: prevent images in issue description from floating into poker elements
 
 ### [3.12.0] - 2019-04-05